move InputSourceSelectionSection to its own file
